# MessageReceiverNotifier non-blocking start