--- a/src/Tests/Seam/When_message_pump_is_failing_to_receive_messages.cs
+++ b/src/Tests/Seam/When_message_pump_is_failing_to_receive_messages.cs
@@ -146,9 +146,12 @@
                 onError = func;
             }
 
+            public void OnCritical(Action<Exception> action)
+            {
+            }
+
             public void OnProcessingFailure(Func<ErrorContext, Task<ErrorHandleResult>> func)
             {
-
             }
 
             public Func<Exception, Task> onError;