--- a/src/Transport/Connectivity/MessageReceiverLifeCycleManager.cs
+++ b/src/Transport/Connectivity/MessageReceiverLifeCycleManager.cs
@@ -1,6 +1,8 @@
 namespace NServiceBus.Transport.AzureServiceBus
 {
     using System.Collections.Concurrent;
+    using System.Threading;
+    using System.Threading.Tasks;
     using Settings;
 
     class MessageReceiverLifeCycleManager
@@ -11,43 +13,55 @@ namespace NServiceBus.Transport.AzureServiceBus
             numberOfReceiversPerEntity = settings.Get<int>(WellKnownConfigurationKeys.Connectivity.NumberOfClientsPerEntity);
         }
 
-        public IMessageReceiverInternal Get(string entityPath, string namespaceAlias)
+        public async Task<IMessageReceiverInternal> Get(string entityPath, string namespaceAlias)
         {
-            var buffer = MessageReceivers.GetOrAdd(entityPath + namespaceAlias, s =>
+            var buffer = await MessageReceivers.GetOrAdd(entityPath + namespaceAlias, async s =>
             {
                 var b = new CircularBuffer<EntityClientEntry>(numberOfReceiversPerEntity);
                 for (var i = 0; i < numberOfReceiversPerEntity; i++)
                 {
-                    var e = new EntityClientEntry();
-                    e.ClientEntity = receiveFactory.Create(entityPath, namespaceAlias).GetAwaiter().GetResult();
+                    var e = new EntityClientEntry
+                    {
+                        ClientEntity = await receiveFactory.Create(entityPath, namespaceAlias)
+                            .ConfigureAwait(false)
+                    };
                     b.Put(e);
                 }
                 return b;
-            });
+            }).ConfigureAwait(false);
 
             var entry = buffer.Get();
 
-            if (entry.ClientEntity.IsClosed)
+            if (!entry.ClientEntity.IsClosed)
             {
-                lock (entry.Mutex)
-                {
-                    if (entry.ClientEntity.IsClosed)
-                    {
-                        entry.ClientEntity = receiveFactory.Create(entityPath, namespaceAlias).GetAwaiter().GetResult();
-                    }
-                }
+                return entry.ClientEntity;
             }
-
-            return entry.ClientEntity;
+            
+            try
+            {
+                await entry.Semaphore.WaitAsync()
+                    .ConfigureAwait(false);
+                
+                if (entry.ClientEntity.IsClosed)
+                {
+                    entry.ClientEntity = await receiveFactory.Create(entityPath, namespaceAlias)
+                        .ConfigureAwait(false);
+                }
+                return entry.ClientEntity;
+            }
+            finally
+            {
+                entry.Semaphore.Release();
+            }
         }
 
         ICreateMessageReceiversInternal receiveFactory;
         int numberOfReceiversPerEntity;
-        ConcurrentDictionary<string, CircularBuffer<EntityClientEntry>> MessageReceivers = new ConcurrentDictionary<string, CircularBuffer<EntityClientEntry>>();
+        ConcurrentDictionary<string, Task<CircularBuffer<EntityClientEntry>>> MessageReceivers = new ConcurrentDictionary<string, Task<CircularBuffer<EntityClientEntry>>>();
 
         class EntityClientEntry
         {
-            internal object Mutex = new object();
+            internal SemaphoreSlim Semaphore = new SemaphoreSlim(1);
             internal IMessageReceiverInternal ClientEntity;
         }
     }