--- a/src/Transport/Connectivity/MessageSenderLifeCycleManager.cs
+++ b/src/Transport/Connectivity/MessageSenderLifeCycleManager.cs
@@ -1,6 +1,8 @@
 namespace NServiceBus.Transport.AzureServiceBus
 {
     using System.Collections.Concurrent;
+    using System.Threading;
+    using System.Threading.Tasks;
     using Settings;
 
     class MessageSenderLifeCycleManager
@@ -11,44 +13,57 @@ namespace NServiceBus.Transport.AzureServiceBus
             numberOfSendersPerEntity = settings.Get<int>(WellKnownConfigurationKeys.Connectivity.NumberOfClientsPerEntity);
         }
 
-        public IMessageSenderInternal Get(string entitypath, string viaEntityPath, string namespaceName)
+        public async Task<IMessageSenderInternal> Get(string entitypath, string viaEntityPath, string namespaceName)
         {
-            var buffer = MessageSenders.GetOrAdd(entitypath + viaEntityPath + namespaceName, s =>
+            var buffer = await MessageSenders.GetOrAdd(entitypath + viaEntityPath + namespaceName, async s =>
             {
                 var b = new CircularBuffer<EntityClientEntry>(numberOfSendersPerEntity);
                 for (var i = 0; i < numberOfSendersPerEntity; i++)
                 {
-                    var e = new EntityClientEntry();
-                    e.ClientEntity = senderFactory.Create(entitypath, viaEntityPath, namespaceName).GetAwaiter().GetResult();
+                    var e = new EntityClientEntry
+                    {
+                        ClientEntity = await senderFactory.Create(entitypath, viaEntityPath, namespaceName)
+                            .ConfigureAwait(false)
+                    };
                     b.Put(e);
                 }
 
                 return b;
-            });
+            }).ConfigureAwait(false);
 
             var entry = buffer.Get();
 
-            if (entry.ClientEntity.IsClosed)
+            if (!entry.ClientEntity.IsClosed)
             {
-                lock (entry.Mutex)
-                {
-                    if (entry.ClientEntity.IsClosed)
-                    {
-                        entry.ClientEntity = senderFactory.Create(entitypath, viaEntityPath, namespaceName).GetAwaiter().GetResult();
-                    }
-                }
+                return entry.ClientEntity;
             }
 
-            return entry.ClientEntity;
+            try
+            {
+                await entry.Semaphore.WaitAsync()
+                    .ConfigureAwait(false);
+                
+                if (entry.ClientEntity.IsClosed)
+                {
+                    entry.ClientEntity = await senderFactory.Create(entitypath, viaEntityPath, namespaceName)
+                        .ConfigureAwait(false);
+                }
+
+                return entry.ClientEntity;
+            }
+            finally
+            {
+                entry.Semaphore.Release();
+            }
         }
 
         ICreateMessageSendersInternal senderFactory;
         int numberOfSendersPerEntity;
-        ConcurrentDictionary<string, CircularBuffer<EntityClientEntry>> MessageSenders = new ConcurrentDictionary<string, CircularBuffer<EntityClientEntry>>();
+        ConcurrentDictionary<string, Task<CircularBuffer<EntityClientEntry>>> MessageSenders = new ConcurrentDictionary<string, Task<CircularBuffer<EntityClientEntry>>>();
 
         class EntityClientEntry
         {
-            internal object Mutex = new object();
+            internal SemaphoreSlim Semaphore = new SemaphoreSlim(1);
             internal IMessageSenderInternal ClientEntity;
         }
     }