@danielmarbach 2017-09-11
@Particular/azure-maintainers please review. No idea why the build is currently failing with the inspect output xml message 

@Particular/deployment-maintainers have you seen this error before? Any hints?