@SeanFeldman 2017-09-11
Build was failing due to inspections executing an outated EAP version of JetBrains tool. Fixed.