@danielmarbach 2017-09-11
@SeanFeldman NRE should be fixed. My bad since there is a race between start and stop that had to be properly addressed